libraries/graphics/scripts/requirements.txt
pillow>=11.1.0
numpy>=2.2.2
imageio>=2.37.0
